+++ web/api.php
@@ -1,58 +1,100 @@
 <?php
 
-require_once __DIR__.'/../silex.phar'; 
+require __DIR__.'/../vendor/autoload.php';
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Serializer\Serializer;
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
 
 use Transport\Entity\Location\Station;
 use Transport\Entity\Location\LocationQuery;
 use Transport\Entity\Location\NearbyQuery;
-use Transport\Entity\Schedule\ConnectionQuery;
+use Transport\Web\ConnectionQueryParser;
+use Transport\Web\LocationQueryParser;
 use Transport\Entity\Schedule\StationBoardQuery;
-
-use Transport\ResultLimit;
+use Transport\Normalizer\FieldsNormalizer;
 
 date_default_timezone_set('Europe/Zurich');
 
 // init
 $app = new Silex\Application();
+
+// default config
 $app['debug'] = true;
+$app['http_cache'] = false;
+$app['buzz.client'] = null;
+$app['monolog.level'] = Monolog\Logger::ERROR;
+$app['xhprof'] = false;
+$app['redis.config'] = false; // array('host' => 'localhost', 'port' => 6379);
+$app['proxy'] = false;
 
+/// load config
+$config = __DIR__.'/../config.php';
+if (stream_resolve_include_path($config)) {
+	include $config;
+}
 
-// autoload
-$app['autoloader']->registerNamespace('Transport', __DIR__.'/../lib');
-$app['autoloader']->registerNamespace('Buzz', __DIR__.'/../vendor/buzz/lib');
-$app['autoloader']->registerNamespace('Predis', __DIR__.'/../vendor/predis/lib');
+// HTTP cache
+if ($app['http_cache']) {
+	$app->register(new Silex\Provider\HttpCacheServiceProvider(), array(
+	    'http_cache.cache_dir' => __DIR__.'/../var/cache/',
+	    'http_cache.options' => array('debug' => $app['debug']),
+	));
+}
 
+// Monolog
+$app->register(new Silex\Provider\MonologServiceProvider(), array(
+    'monolog.logfile' => __DIR__.'/../var/logs/transport.log',
+    'monolog.level' => $app['monolog.level'],
+    'monolog.name' => 'transport',
+));
+$app->before(function (Request $request) use ($app) {
+    $app['monolog']->addInfo('- ' . $request->getClientIp() . ' ' . $request->headers->get('referer') . ' ' . $request->server->get('HTTP_USER_AGENT'));
+});
+
+// XHProf
+if ($app['xhprof']) {
+    xhprof_enable();
+}
+
+// if hosted behind a reverse proxy
+if ($app['proxy']) {
+    Request::trustProxyData();
+}
 
 // create Transport API
-$app['api'] = new Transport\API();
-
+$app['api'] = new Transport\API(new Buzz\Browser($app['buzz.client']));
 
 // allow cross-domain requests, enable cache
 $app->after(function (Request $request, Response $response) {
     $response->headers->set('Access-Control-Allow-Origin', '*');
-    $response->headers->set('Cache-Control', 's-maxage=30');
+    $response->headers->set('Cache-Control', 's-maxage=30, public');
+});
+
+// Serializer
+$app['serializer'] = $app->share(function () use ($app) {
+    $fields = $app['request']->get('fields') ?: array();
+    return new Serializer(array(new FieldsNormalizer($fields)), array('json' => new JsonEncoder()));
 });
 
 
-// count API calls
-$app['redis'] = new Predis\Client(array('host' => 'tetra.redistogo.com', 'port' => 9464, 'password' => '7cd7bdf5a51d601547da3c96d6bae1a2'));
+// statistics
+$redis = null;
 try {
-    $app['redis']->connect();
-    $app->after(function (Request $request, Response $response) use ($app) {
-
-        $date = date('Y-m-d');
-        $key = "stats:calls:$date";
-
-        $app['redis']->incr($key);
-    });
-} catch (Predis\Network\ConnectionException $e) {
-    // ignore connection error
-} catch (Predis\ServerException $e) {
-    // ignore connection error
+    if ($app['redis.config']) {
+        $redis = new Predis\Client($app['redis.config']);
+        $redis->connect();
+    }
+} catch (Exception $e) {
+    $app['monolog']->addError($e->getMessage());
+    $redis = null;
 }
+$app['stats'] = new Transport\Statistics($redis);
+$app->after(function (Request $request, Response $response) use ($app) {
+    $app['stats']->call();
+    $app['stats']->resource($request->getPathInfo());
+});
 
 
 // index
@@ -90,48 +132,21 @@ $app->get('/v1/locations', function(Request $request) use ($app) {
         $stations = $app['api']->findLocations($query);
     }
 
-    return $app->json(array('stations' => $stations));
+    $result = array('stations' => $stations);
+
+    $json = $app['serializer']->serialize((object) $result, 'json');
+    return new Response($json, 200, array('Content-Type' => 'application/json'));
 });
 
 
 // connections
 $app->get('/v1/connections', function(Request $request) use ($app) {
-    // validate
-    $from = $request->get('from');
-    $to = $request->get('to');
-    if (is_array($request->get('via'))) {
-        $via = $request->get('via');
-        if (count($via) > 5) {
-            return new Response('Invalid via count (max 5 allowed)', 400);
-        }
-    } else if ($request->get('via')) {
-        $via = array($request->get('via'));
-    } else {
-        $via = array();
-    }
-    $date = $request->get('date') ?: null;
-    $time = $request->get('time') ?: null;
-    $isArrivalTime = $request->get('isArrivalTime') ?: null;
-    $limit = $request->get('limit') ?: 4;
-    $transportations = $request->get('transportations');
-    $direct = $request->get('direct');
-    $sleeper = $request->get('sleeper');
-    $couchette = $request->get('chouchette');
-    $bike = $request->get('bike');
 
-    ResultLimit::setFields($request->get('fields') ?: array());
-
-    if ($limit > 6) {
-        return new Response('Maximal value of argument `limit` is 6.', 400);
-    }
+    $query = LocationQueryParser::create($request);
 
     // get stations
-    $stations = array('from' => array(), 'to' => array(), 'via' => array());
-    if ($from && $to) {
-        $query = new LocationQuery(array('from' => $from, 'to' => $to, 'via' => $via));
-        $stations = $app['api']->findLocations($query);
-    }
-    
+    $stations = $app['api']->findLocations($query);
+
     // get connections
     $connections = array();
     $from = reset($stations['from']) ?: null;
@@ -144,65 +159,35 @@ $app->get('/v1/connections', function(Request $request) use ($app) {
     }
 
     if ($from && $to) {
-        $query = new ConnectionQuery($from, $to, $via, $date, $time);
-        $query->forwardCount = $limit;
-        $query->backwardCount = 0;
-        if ($isArrivalTime !== null) {
-            switch ($isArrivalTime) {
-                case 0:
-                case "false":
-                    $query->isArrivalTime = false;
-                    break;
-                case 1:
-                case "true":
-                    $query->isArrivalTime = true;
-                    break;
-                default:
-                    //wrong parameter value
-                    break;
-            }
+        $app['stats']->station($from);
+        $app['stats']->station($to);
+
+        $query = ConnectionQueryParser::create($request, $from, $to, $via);
+
+        $errors = ConnectionQueryParser::validate($query);
+        if ($errors) {
+            return $app->json(array('errors' => $errors), 400);
         }
-        if ($transportations) {
-            $query->transportations = $transportations;
-        }
-        if ($direct) {
-            $query->direct = $direct;
-        }
-        if ($sleeper) {
-            $query->sleeper = $sleeper;
-        }
-        if ($couchette) {
-            $query->couchette = $couchette;
-        }
-        if ($bike) {
-            $query->bike = $bike;
-        }
-        $connections = $app['api']->findConnections($query, 'connections');
+
+        $connections = $app['api']->findConnections($query);
     }
-    $result = array('connections' => $connections);
-    if (ResultLimit::isFieldSet('from')) {
-        $result = array_merge($result,array('from' => $from));   
-    }
-    if (ResultLimit::isFieldSet('to')) {
-        $result = array_merge($result,array('to' => $to));   
-    }
-    if (ResultLimit::isFieldSet('stations')) {
-        $result = array_merge($result,array('stations' => $stations));   
-    }
-    return $app->json($result);
+
+    $result = array(
+        'connections' => $connections,
+        'from' => $from,
+        'to' => $to,
+        'stations' => $stations,
+    );
+
+    $json = $app['serializer']->serialize((object) $result, 'json');
+    return new Response($json, 200, array('Content-Type' => 'application/json'));
 });
 
 
 // station board
 $app->get('/v1/stationboard', function(Request $request) use ($app) {
 
-    $station = null;
     $stationboard = array();
-
-    $id = $request->get('id');
-    if ($id) {
-        $station = new Station($id);
-    }
 
     $limit = $request->get('limit');
     if ($limit > 420) {
@@ -210,34 +195,54 @@ $app->get('/v1/stationboard', function(Request $request) use ($app) {
     }
 
     $date = $request->get('date');
+    if (!$date) {
+        $date = $request->get('datetime');
+    }
     if ($date) {
         $date = new DateTime($date, new DateTimeZone('Europe/Zurich'));
     }
 
     $transportations = $request->get('transportations');
-    
-    ResultLimit::setFields($request->get('fields') ?: array());
 
-    if (!$station) {
+    $station = $request->get('station') ?: $request->get('id');
 
-        $station = $request->get('station');
+    $query = new LocationQuery($station);
+    $stations = $app['api']->findLocations($query);
+    $station = reset($stations);
 
-        $query = new LocationQuery($station);
-        $stations = $app['api']->findLocations($query);
-        $station = reset($stations);
-    }
+    if ($station instanceof Station) {
+        $app['stats']->station($station);
 
-    if ($station) {
         $query = new StationBoardQuery($station, $date);
         if ($transportations) {
             $query->transportations = $transportations;
         }
         $query->maxJourneys = $limit;
-        $stationboard = $app['api']->getStationBoard($query, 'stationboard');
+        $stationboard = $app['api']->getStationBoard($query);
     }
-    return $app->json(array('stationboard' => $stationboard));
+
+    $result = array('station' => $station, 'stationboard' => $stationboard);
+
+    $json = $app['serializer']->serialize((object) $result, 'json');
+    return new Response($json, 200, array('Content-Type' => 'application/json'));
 });
 
 
 // run
-$app->run(); 
+if ($app['http_cache']) {
+    $app['http_cache']->run();
+} else {
+	$app->run();
+}
+
+// save XHProf run
+if ($app['xhprof']) {
+
+    $data = xhprof_disable();
+
+    include_once __DIR__.'/../vendor/facebook/xhprof/xhprof_lib/utils/xhprof_lib.php';
+    include_once __DIR__.'/../vendor/facebook/xhprof/xhprof_lib/utils/xhprof_runs.php';
+
+    $xhprof = new XHProfRuns_Default(__DIR__.'/../var/xhprof');
+    $run_id = $xhprof->save_run($data, 'transport');
+}